pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "context-agent"
version = "0.1.0"
description = "A context proxy built on openJiuwen for unified, extensible, and governable context management"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
dependencies = [
    "openjiuwen",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "redis[hiredis]>=5.0.0",
    "chromadb>=0.5.0",
    "structlog>=24.1.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
]

[project.optional-dependencies]
pulsar = ["pulsar-client>=3.4.0"]
neo4j = ["neo4j>=5.20.0"]
milvus = ["pymilvus>=2.4.0"]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-benchmark>=4.0.0",
    "testcontainers[redis,compose]>=4.7.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-redis>=4.6.0",
]

[project.scripts]
context-agent = "context_agent.api.http_handler:main"

[tool.hatch.build.targets.wheel]
packages = ["context_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--strict-markers -q"
markers = [
    "unit: unit tests (no external deps)",
    "integration: integration tests (requires Redis/Chroma)",
    "performance: performance regression tests",
]

[tool.coverage.run]
source = ["context_agent"]
omit = ["context_agent/__main__.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
